+++ src/complex/complex.lib.js
@@ -1,2 +1,3 @@
 export * as rect from "./rect/rect.lib.js";
+export * from "./rect/rect.lib.js"; //Default form is rectangular
 export * as polar from "./polar/polar.lib.js";