+++ src/random/random.lib.js
@@ -1,3 +1,5 @@
-export * from "./random.src.js";
-//export * as map from "./map/map.lib.js";
-//export * as noise from "./noise/noise.lib.js";
+export * from "./sequence.src.js";
+export * from "./map.src.js";
+export * from "./perlin1d.src.js";
+export * from "./perlin2d.src.js";
+export * as debug from "./random.debug.js";