+++ src/vector/polar/polar.src.js
@@ -8,17 +8,22 @@ function mag(vec) {
 	return abs(vec[0]);
 }
 
-function smult2(vec, k, target = new Float64Array(2)) {
+function scale2(vec, k, target = new Float64Array(2)) {
 	target[0] = vec[0] * k;
 	target[1] = mod(vec[1], TWO_PI);
+	return target;
 }
 
 function normalize2(vec, target = new Float64Array(2)) {
+	if (vec[0] === 0) {
+		return undefined;
+	}
 	target[0] = 1;
 	target[1] = mod(vec[1], TWO_PI);
+	return target;
 }
 
-function rect2(vec, target = new Float64Array(2)) {
+function toRect2(vec, target = new Float64Array(2)) {
 	const r = vec[0];
 	const theta = vec[1];
 	target[0] = r * cos(theta);
@@ -29,9 +34,9 @@ function rect2(vec, target = new Float64Array(2)) {
 // Freeze exports
 Object.freeze(dot2);
 Object.freeze(mag);
-Object.freeze(smult2);
+Object.freeze(scale2);
 Object.freeze(normalize2);
-Object.freeze(rect2);
+Object.freeze(toRect2);
 
 // Export
-export {dot2, mag, smult2, normalize2, rect2}
+export {dot2, mag, scale2, normalize2, toRect2}