+++ test/unit/mathicall.tests.js
@@ -1,4 +1,0 @@
-import "./src/array.tests.js";
-import "./src/random.tests.js";
-import "./src/matrix.tests.js";
-import "./src/integer.tests.js";